customanchor: fix the package size guard on 32-bit builds
len() is a 32-bit int on arm, so comparing it against the untyped
math.MaxUint32 constant fails to compile. Widen to uint64 first.

+++ custom_anchor_package.go
@@ -1277,7 +1277,7 @@ func (p *CustomAnchorTransferPackage) MarshalBinary() ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	if len(jsonBytes) > math.MaxUint32 {
+	if uint64(len(jsonBytes)) > math.MaxUint32 {
 		return nil, fmt.Errorf("custom anchor package is too large")
 	}
